+++ app/build.gradle
@@ -24,7 +24,7 @@ android {
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "jp.co.yumemi.android.codecheck.CustomTestRunner"
         buildConfigField("String", "GITHUB_ACCESS_TOKEN", "\"" + localProperties['GITHUB_ACCESS_TOKEN'] + "\"")
         buildConfigField "String", "GITHUB_API_DOMAIN", "\"https://api.github.com\""
     }
@@ -44,6 +44,16 @@ android {
     }
     buildFeatures {
         viewBinding true
+    }
+    // Tests can be Robolectric or instrumented tests
+    sourceSets {
+        String sharedTestDir = 'src/sharedTest/kotlin'
+        test {
+            java.srcDir sharedTestDir
+        }
+        androidTest {
+            java.srcDir sharedTestDir
+        }
     }
 }
 
@@ -66,12 +76,17 @@ dependencies {
 
     // Test
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
+    androidTestImplementation 'androidx.test.ext:junit-ktx:1.1.4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.5.2"
 
     // DI
     implementation "com.google.dagger:hilt-android:$dagger_hilt_version"
     kapt "com.google.dagger:hilt-compiler:$dagger_hilt_version"
+    // DI: test
+    testImplementation "com.google.dagger:hilt-android-testing:$dagger_hilt_version"
+    androidTestImplementation "com.google.dagger:hilt-android-testing:$dagger_hilt_version"
+    kaptAndroidTest "com.google.dagger:hilt-compiler:$dagger_hilt_version"
 
     // json
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:1.4.0"
