+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/CustomTestRunner.kt
@@ -1,0 +1,16 @@
+package jp.co.yumemi.android.codecheck
+
+import android.app.Application
+import android.content.Context
+import androidx.test.runner.AndroidJUnitRunner
+import dagger.hilt.android.testing.HiltTestApplication
+
+class CustomTestRunner : AndroidJUnitRunner() {
+    override fun newApplication(
+        cl: ClassLoader?,
+        className: String?,
+        context: Context?
+    ): Application {
+        return super.newApplication(cl, HiltTestApplication::class.java.name, context)
+    }
+}
