+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/package-info.kt
@@ -1,0 +1,5 @@
+package jp.co.yumemi.android.codecheck
+
+/*
+Tests for Composable functions that use the android library are placed here.
+ */
