--- app/src/main/kotlin/jp/co/yumemi/android/code_check/MyApp.kt
@@ -1,7 +1,0 @@
-package jp.co.yumemi.android.code_check
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MyApp : Application()
