--- app/src/main/kotlin/jp/co/yumemi/android/code_check/di/ApiModule.kt
@@ -1,21 +1,0 @@
-package jp.co.yumemi.android.code_check.di
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import io.ktor.client.*
-import io.ktor.client.engine.android.*
-import io.ktor.client.request.*
-import io.ktor.http.*
-import jp.co.yumemi.android.code_check.data.api.GithubApi
-
-import javax.inject.Singleton
-
-@InstallIn(SingletonComponent::class)
-@Module
-object ApiModule {
-    @Singleton
-    @Provides
-    fun provideGithubService(): GithubApi = GithubApi()
-}
