--- app/src/main/kotlin/jp/co/yumemi/android/code_check/di/DispatcherModule.kt
@@ -1,16 +1,0 @@
-package jp.co.yumemi.android.code_check.di
-
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-@Module
-@InstallIn(SingletonComponent::class)
-object DispatcherModule {
-    @IoDispatcher
-    @Provides
-    fun providesIoDispatcher(): CoroutineDispatcher = Dispatchers.IO
-}
