--- app/src/main/kotlin/jp/co/yumemi/android/code_check/di/Dispatchers.kt
@@ -1,7 +1,0 @@
-package jp.co.yumemi.android.code_check.di
-
-import javax.inject.Qualifier
-
-@Retention(AnnotationRetention.BINARY)
-@Qualifier
-annotation class IoDispatcher
