+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/MyApp.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.codecheck
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApp : Application()
