+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/TopActivity.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright © 2021 YUMEMI Inc. All rights reserved.
+ */
+package jp.co.yumemi.android.codecheck
+
+import android.os.Bundle
+import android.os.PersistableBundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.WindowCompat
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class TopActivity : AppCompatActivity(R.layout.activity_top) {
+    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
+        WindowCompat.setDecorFitsSystemWindows(window, false)
+        super.onCreate(savedInstanceState, persistentState)
+    }
+}
