+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/data/api/response/GithubErrorResponse.kt
@@ -1,0 +1,4 @@
+package jp.co.yumemi.android.codecheck.data.api.response
+
+@kotlinx.serialization.Serializable
+data class GithubErrorResponse(val documentation_url: String, val message: String)
