+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/data/repository/GithubRepoRepository.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.codecheck.data.repository
+
+import jp.co.yumemi.android.codecheck.model.RepositorySummary
+
+interface GithubRepoRepository {
+    suspend fun searchRepositories(query: String): List<RepositorySummary>
+}
