+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/di/ApiModule.kt
@@ -1,0 +1,21 @@
+package jp.co.yumemi.android.codecheck.di
+
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import io.ktor.client.*
+import io.ktor.client.engine.android.*
+import io.ktor.client.request.*
+import io.ktor.http.*
+import jp.co.yumemi.android.codecheck.data.api.GithubApi
+
+import javax.inject.Singleton
+
+@InstallIn(SingletonComponent::class)
+@Module
+object ApiModule {
+    @Singleton
+    @Provides
+    fun provideGithubService(): GithubApi = GithubApi()
+}
