+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/ui/repositories/RepositoriesViewModel.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright © 2021 YUMEMI Inc. All rights reserved.
+ */
+package jp.co.yumemi.android.codecheck.ui.repositories
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import dagger.hilt.android.lifecycle.HiltViewModel
+import jp.co.yumemi.android.codecheck.data.repository.GithubRepoRepository
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.*
+import javax.inject.Inject
+
+@HiltViewModel
+class RepositoriesViewModel @Inject constructor(
+    private val repository: GithubRepoRepository,
+) : ViewModel() {
+
+    private val coroutineExceptionHandler = CoroutineExceptionHandler { _, throwable ->
+        _uiState.value = UiState.Error(throwable)
+    }
+
+    private val _uiState: MutableStateFlow<UiState> = MutableStateFlow(UiState.Initial)
+    val uiState: StateFlow<UiState> = _uiState.asStateFlow()
+
+    fun searchResults(query: String) {
+        viewModelScope.launch(coroutineExceptionHandler) {
+            _uiState.value = UiState.Loading
+            val items = repository.searchRepositories(query)
+            _uiState.value = UiState.Data(items)
+        }
+    }
+}
