+++ app/src/main/res/layout/fragment_repositories.xml
@@ -72,6 +72,28 @@
             android:layout_height="match_parent"
             android:layout_gravity="center"
             android:indeterminate="true" />
+
+        <LinearLayout
+            android:id="@+id/errorLayout"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:gravity="center"
+            android:orientation="vertical">
+
+            <TextView
+                android:id="@+id/errorMessage"
+                style="@style/TextAppearance.Material3.BodyLarge"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_margin="20dp"
+                tools:text="エラーメッセージをここに表示する" />
+
+            <Button
+                android:id="@+id/buttonReload"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/reload" />
+        </LinearLayout>
     </FrameLayout>
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
