+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/repositoriesFragment"
-        android:name="jp.co.yumemi.android.code_check.ui.repositories.RepositoriesFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.repositories.RepositoriesFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_repositories">
         <action
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/repositoryFragment"
-        android:name="jp.co.yumemi.android.code_check.ui.repository.RepositoryFragment"
+        android:name="jp.co.yumemi.android.codecheck.ui.repository.RepositoryFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_repository">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.model.RepositorySummary" />
+            app:argType="jp.co.yumemi.android.codecheck.model.RepositorySummary" />
     </fragment>
 
 </navigation>
