+++ app/src/test/kotlin/jp/co/yumemi/android/codecheck/package-info.kt
@@ -1,0 +1,5 @@
+package jp.co.yumemi.android.codecheck
+
+/*
+ViewModel, business logic test code is placed here.
+ */
